feat(replay): Do not capture replays < 5 seconds (GA) (#8277)
