+++ packages/browser-integration-tests/suites/replay/bufferMode/test.ts
@@ -25,7 +25,6 @@ sentryTest(
     let errorEventId: string | undefined;
     const reqPromise0 = waitForReplayRequest(page, 0);
     const reqPromise1 = waitForReplayRequest(page, 1);
-    const reqPromise2 = waitForReplayRequest(page, 2);
     const reqErrorPromise = waitForErrorRequest(page);
 
     await page.route('https://dsn.ingest.sentry.io/**/*', route => {
@@ -101,17 +100,13 @@ sentryTest(
 
     // Switches to session mode and then goes to background
     const req1 = await reqPromise1;
-    const req2 = await reqPromise2;
-    expect(callsToSentry).toBeGreaterThanOrEqual(5);
+    expect(callsToSentry).toBeGreaterThanOrEqual(4);
 
     const event0 = getReplayEvent(req0);
     const content0 = getReplayRecordingContent(req0);
 
     const event1 = getReplayEvent(req1);
     const content1 = getReplayRecordingContent(req1);
-
-    const event2 = getReplayEvent(req2);
-    const content2 = getReplayRecordingContent(req2);
 
     expect(event0).toEqual(
       getExpectedReplayEvent({
@@ -157,17 +152,7 @@ sentryTest(
 
     // From switching to session mode
     expect(content1.fullSnapshots).toHaveLength(1);
-
-    expect(event2).toEqual(
-      getExpectedReplayEvent({
-        replay_type: 'buffer', // although we're in session mode, we still send 'buffer' as replay_type
-        segment_id: 2,
-        urls: [],
-      }),
-    );
-
-    expect(content2.fullSnapshots).toHaveLength(0);
-    expect(content2.breadcrumbs).toEqual(expect.arrayContaining([expectedClickBreadcrumb]));
+    expect(content1.breadcrumbs).toEqual(expect.arrayContaining([expectedClickBreadcrumb]));
   },
 );
