+++ packages/browser-integration-tests/suites/replay/largeMutations/defaultOptions/test.ts
@@ -11,7 +11,6 @@ sentryTest(
     }
 
     const reqPromise0 = waitForReplayRequest(page, 0);
-    const reqPromise0b = waitForReplayRequest(page, 1);
 
     await page.route('https://dsn.ingest.sentry.io/**/*', route => {
       return route.fulfill({
@@ -24,10 +23,7 @@ sentryTest(
     const url = await getLocalTestPath({ testDir: __dirname });
 
     await page.goto(url);
-    await forceFlushReplay();
     const res0 = await reqPromise0;
-    await reqPromise0b;
-    // A second request is sent right after initial snapshot, we want to wait for that to settle before we continue
 
     const reqPromise1 = waitForReplayRequest(page);
