+++ packages/browser-integration-tests/suites/replay/slowClick/mutation/test.ts
@@ -3,7 +3,7 @@ import { expect } from '@playwright/test';
 import { sentryTest } from '../../../../utils/fixtures';
 import { getCustomRecordingEvents, shouldSkipReplayTest, waitForReplayRequest } from '../../../../utils/replayHelpers';
 
-sentryTest('mutation after threshold results in slow click', async ({ getLocalTestUrl, page }) => {
+sentryTest('mutation after threshold results in slow click', async ({ forceFlushReplay, getLocalTestUrl, page }) => {
   if (shouldSkipReplayTest()) {
     sentryTest.skip();
   }
@@ -21,6 +21,7 @@ sentryTest('mutation after threshold results in slow click', async ({ getLocalTe
   const url = await getLocalTestUrl({ testDir: __dirname });
 
   await page.goto(url);
+  await forceFlushReplay();
   await reqPromise0;
 
   const reqPromise1 = waitForReplayRequest(page, (event, res) => {
@@ -125,59 +126,63 @@ sentryTest('multiple clicks are counted', async ({ getLocalTestUrl, page }) => {
   expect(slowClickBreadcrumbs[0]?.data?.timeAfterClickMs).toBeLessThan(3100);
 });
 
-sentryTest('immediate mutation does not trigger slow click', async ({ browserName, getLocalTestUrl, page }) => {
-  // This test seems to only be flakey on firefox
-  if (shouldSkipReplayTest() || ['firefox'].includes(browserName)) {
-    sentryTest.skip();
-  }
+sentryTest(
+  'immediate mutation does not trigger slow click',
+  async ({ forceFlushReplay, browserName, getLocalTestUrl, page }) => {
+    // This test seems to only be flakey on firefox
+    if (shouldSkipReplayTest() || ['firefox'].includes(browserName)) {
+      sentryTest.skip();
+    }
 
-  const reqPromise0 = waitForReplayRequest(page, 0);
+    const reqPromise0 = waitForReplayRequest(page, 0);
 
-  await page.route('https://dsn.ingest.sentry.io/**/*', route => {
-    return route.fulfill({
-      status: 200,
-      contentType: 'application/json',
-      body: JSON.stringify({ id: 'test-id' }),
+    await page.route('https://dsn.ingest.sentry.io/**/*', route => {
+      return route.fulfill({
+        status: 200,
+        contentType: 'application/json',
+        body: JSON.stringify({ id: 'test-id' }),
+      });
     });
-  });
 
-  const url = await getLocalTestUrl({ testDir: __dirname });
+    const url = await getLocalTestUrl({ testDir: __dirname });
 
-  await page.goto(url);
-  await reqPromise0;
+    await page.goto(url);
+    await forceFlushReplay();
+    await reqPromise0;
 
-  const reqPromise1 = waitForReplayRequest(page, (event, res) => {
-    const { breadcrumbs } = getCustomRecordingEvents(res);
+    const reqPromise1 = waitForReplayRequest(page, (event, res) => {
+      const { breadcrumbs } = getCustomRecordingEvents(res);
 
-    return breadcrumbs.some(breadcrumb => breadcrumb.category === 'ui.click');
-  });
+      return breadcrumbs.some(breadcrumb => breadcrumb.category === 'ui.click');
+    });
 
-  await page.click('#mutationButtonImmediately');
+    await page.click('#mutationButtonImmediately');
 
-  const { breadcrumbs } = getCustomRecordingEvents(await reqPromise1);
+    const { breadcrumbs } = getCustomRecordingEvents(await reqPromise1);
 
-  expect(breadcrumbs).toEqual([
-    {
-      category: 'ui.click',
-      data: {
-        node: {
-          attributes: {
-            id: 'mutationButtonImmediately',
+    expect(breadcrumbs).toEqual([
+      {
+        category: 'ui.click',
+        data: {
+          node: {
+            attributes: {
+              id: 'mutationButtonImmediately',
+            },
+            id: expect.any(Number),
+            tagName: 'button',
+            textContent: '******* ******** ***********',
           },
-          id: expect.any(Number),
-          tagName: 'button',
-          textContent: '******* ******** ***********',
+          nodeId: expect.any(Number),
         },
-        nodeId: expect.any(Number),
+        message: 'body > button#mutationButtonImmediately',
+        timestamp: expect.any(Number),
+        type: 'default',
       },
-      message: 'body > button#mutationButtonImmediately',
-      timestamp: expect.any(Number),
-      type: 'default',
-    },
-  ]);
-});
+    ]);
+  },
+);
 
-sentryTest('inline click handler does not trigger slow click', async ({ getLocalTestUrl, page }) => {
+sentryTest('inline click handler does not trigger slow click', async ({ forceFlushReplay, getLocalTestUrl, page }) => {
   if (shouldSkipReplayTest()) {
     sentryTest.skip();
   }
@@ -195,6 +200,7 @@ sentryTest('inline click handler does not trigger slow click', async ({ getLocal
   const url = await getLocalTestUrl({ testDir: __dirname });
 
   await page.goto(url);
+  await forceFlushReplay();
   await reqPromise0;
 
   const reqPromise1 = waitForReplayRequest(page, (event, res) => {
