+++ packages/e2e-tests/test-applications/standard-frontend-react/tests/fixtures/ReplayRecordingData.ts
@@ -96,26 +96,6 @@ export const ReplayRecordingData = [
       timestamp: expect.any(Number),
     },
     {
-      type: 5,
-      timestamp: expect.any(Number),
-      data: {
-        tag: 'performanceSpan',
-        payload: {
-          op: 'memory',
-          description: 'memory',
-          startTimestamp: expect.any(Number),
-          endTimestamp: expect.any(Number),
-          data: {
-            memory: {
-              jsHeapSizeLimit: expect.any(Number),
-              totalJSHeapSize: expect.any(Number),
-              usedJSHeapSize: expect.any(Number),
-            },
-          },
-        },
-      },
-    },
-    {
       type: 3,
       data: {
         source: 0,
@@ -155,8 +135,6 @@ export const ReplayRecordingData = [
       data: { source: 5, text: 'Capture Exception', isChecked: false, id: 16 },
       timestamp: expect.any(Number),
     },
-  ],
-  [
     {
       type: 5,
       timestamp: expect.any(Number),
